+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#751](https://github.com/babylonlabs-io/finality-provider/pull/751) feat: remove failing msg from batch
+
 ## v2.0.0
 
 ### Improvements